pyproject.toml
[project]
name = "web-app-company-django"
version = "0.1.0"
description = ""
authors = [
    {name = "ragecodemode",email = "nikitaloskutov18@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "pydantic[email] (>=2.11.1,<3.0.0)",
    "django-ninja (>=1.3.0,<2.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "jwt (>=1.3.1,<2.0.0)",
]


[tool.poetry.dependencies]
python = ">=3.11,<4"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
